refactor profile features
