+++ lib/features/my_profile/presentation/views/widgets/my_profile/show_user_data.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:looqma/core/utils/app_styles.dart';
 import 'package:looqma/features/my_profile/data/models/user_profile_response_model.dart';
-import 'package:looqma/features/my_profile/presentation/views/widgets/user_data_item.dart';
-import 'package:looqma/features/my_profile/presentation/views/widgets/user_image_section.dart';
+import 'package:looqma/features/my_profile/presentation/views/widgets/my_profile/user_data_item.dart';
+import 'package:looqma/features/my_profile/presentation/views/widgets/my_profile/user_image_section.dart';
 
 class ShowUserData extends StatelessWidget {
   const ShowUserData({
